+++ bptree/insert.go
@@ -1,0 +1,243 @@
+package bptree
+
+import "errors"
+
+// Insert - implement duplicate key insertion functionality
+// create new node w/o parent pointer (point to the left)
+func (t *Tree) Insert(key int, value []byte) error {
+	var pointer *Record
+	var leaf *Node
+
+	// edit
+	if _, err := t.Find(key, false); err == nil {
+		return errors.New("key already exists")
+	}
+
+	pointer, err := makeRecord(value)
+	if err != nil {
+		return err
+	}
+
+	if t.Root == nil {
+		return t.startNewTree(key, pointer)
+	}
+
+	leaf = t.findLeaf(key, false)
+
+	if leaf.NumKeys < order-1 {
+		insertIntoLeaf(leaf, key, pointer)
+		return nil
+	}
+
+	return t.insertIntoLeafAfterSplitting(leaf, key, pointer)
+}
+
+// implement binsearch
+func getLeftIndex(parent, left *Node) int {
+	left_index := 0
+	for left_index <= parent.NumKeys && parent.Pointers[left_index] != left {
+		left_index += 1
+	}
+	return left_index
+}
+
+// implement binsearch
+func insertIntoLeaf(leaf *Node, key int, pointer *Record) {
+	var i, insertion_point int
+
+	for insertion_point < leaf.NumKeys && leaf.Keys[insertion_point] < key {
+		insertion_point += 1
+	}
+
+	for i = leaf.NumKeys; i > insertion_point; i-- {
+		leaf.Keys[i] = leaf.Keys[i-1]
+		leaf.Pointers[i] = leaf.Pointers[i-1]
+	}
+	leaf.Keys[insertion_point] = key
+	leaf.Pointers[insertion_point] = pointer
+	leaf.NumKeys += 1
+	return
+}
+
+// implement binsearch
+func (t *Tree) insertIntoLeafAfterSplitting(leaf *Node, key int, pointer *Record) error {
+	var new_leaf *Node
+	var insertion_index, split, new_key, i, j int
+	var err error
+
+	new_leaf, err = makeLeaf()
+	if err != nil {
+		return nil
+	}
+
+	temp_keys := make([]int, order)
+	if temp_keys == nil {
+		return errors.New("Error: Temporary keys array.")
+	}
+
+	temp_pointers := make([]interface{}, order)
+	if temp_pointers == nil {
+		return errors.New("Error: Temporary pointers array.")
+	}
+
+	for insertion_index < order-1 && leaf.Keys[insertion_index] < key {
+		insertion_index += 1
+	}
+
+	for i = 0; i < leaf.NumKeys; i++ {
+		if j == insertion_index {
+			j += 1
+		}
+		temp_keys[j] = leaf.Keys[i]
+		temp_pointers[j] = leaf.Pointers[i]
+		j += 1
+	}
+
+	temp_keys[insertion_index] = key
+	temp_pointers[insertion_index] = pointer
+
+	leaf.NumKeys = 0
+
+	split = cut(order - 1)
+
+	for i = 0; i < split; i++ {
+		leaf.Pointers[i] = temp_pointers[i]
+		leaf.Keys[i] = temp_keys[i]
+		leaf.NumKeys += 1
+	}
+
+	j = 0
+	for i = split; i < order; i++ {
+		new_leaf.Pointers[j] = temp_pointers[i]
+		new_leaf.Keys[j] = temp_keys[i]
+		new_leaf.NumKeys += 1
+		j += 1
+	}
+
+	new_leaf.Pointers[order-1] = leaf.Pointers[order-1]
+	leaf.Pointers[order-1] = new_leaf
+
+	for i = leaf.NumKeys; i < order-1; i++ {
+		leaf.Pointers[i] = nil
+	}
+	for i = new_leaf.NumKeys; i < order-1; i++ {
+		new_leaf.Pointers[i] = nil
+	}
+
+	// point to the left for dup keys
+	new_leaf.Parent = leaf.Parent
+	new_key = new_leaf.Keys[0]
+
+	return t.insertIntoParent(leaf, new_key, new_leaf)
+}
+
+func insertIntoNode(n *Node, left_index, key int, right *Node) {
+	var i int
+	for i = n.NumKeys; i > left_index; i-- {
+		n.Pointers[i+1] = n.Pointers[i]
+		n.Keys[i] = n.Keys[i-1]
+	}
+	n.Pointers[left_index+1] = right
+	n.Keys[left_index] = key
+	n.NumKeys += 1
+}
+
+// implement binsearch
+func (t *Tree) insertIntoNodeAfterSplitting(old_node *Node, left_index, key int, right *Node) error {
+	var i, j, split, k_prime int
+	var new_node, child *Node
+	var temp_keys []int
+	var temp_pointers []interface{}
+	var err error
+
+	temp_pointers = make([]interface{}, order+1)
+	if temp_pointers == nil {
+		return errors.New("Error: Temporary pointers array for splitting nodes.")
+	}
+
+	temp_keys = make([]int, order)
+	if temp_keys == nil {
+		return errors.New("Error: Temporary keys array for splitting nodes.")
+	}
+
+	for i = 0; i < old_node.NumKeys+1; i++ {
+		if j == left_index+1 {
+			j += 1
+		}
+		temp_pointers[j] = old_node.Pointers[i]
+		j += 1
+	}
+
+	j = 0
+	for i = 0; i < old_node.NumKeys; i++ {
+		if j == left_index {
+			j += 1
+		}
+		temp_keys[j] = old_node.Keys[i]
+		j += 1
+	}
+
+	temp_pointers[left_index+1] = right
+	temp_keys[left_index] = key
+
+	split = cut(order)
+	new_node, err = makeNode()
+	if err != nil {
+		return err
+	}
+	old_node.NumKeys = 0
+	for i = 0; i < split-1; i++ {
+		old_node.Pointers[i] = temp_pointers[i]
+		old_node.Keys[i] = temp_keys[i]
+		old_node.NumKeys += 1
+	}
+	old_node.Pointers[i] = temp_pointers[i]
+	k_prime = temp_keys[split-1]
+	j = 0
+	for i += 1; i < order; i++ {
+		new_node.Pointers[j] = temp_pointers[i]
+		new_node.Keys[j] = temp_keys[i]
+		new_node.NumKeys += 1
+		j += 1
+	}
+	new_node.Pointers[j] = temp_pointers[i]
+	new_node.Parent = old_node.Parent
+	for i = 0; i <= new_node.NumKeys; i++ {
+		child, _ = new_node.Pointers[i].(*Node)
+		child.Parent = new_node
+	}
+
+	return t.insertIntoParent(old_node, k_prime, new_node)
+}
+
+func (t *Tree) insertIntoParent(left *Node, key int, right *Node) error {
+	var left_index int
+	parent := left.Parent
+
+	if parent == nil {
+		return t.insertIntoNewRoot(left, key, right)
+	}
+	left_index = getLeftIndex(parent, left)
+
+	if parent.NumKeys < order-1 {
+		insertIntoNode(parent, left_index, key, right)
+		return nil
+	}
+
+	return t.insertIntoNodeAfterSplitting(parent, left_index, key, right)
+}
+
+func (t *Tree) insertIntoNewRoot(left *Node, key int, right *Node) error {
+	t.Root, err = makeNode()
+	if err != nil {
+		return err
+	}
+	t.Root.Keys[0] = key
+	t.Root.Pointers[0] = left
+	t.Root.Pointers[1] = right
+	t.Root.NumKeys += 1
+	t.Root.Parent = nil
+	left.Parent = t.Root
+	right.Parent = t.Root
+	return nil
+}
