+++ main.go
@@ -8,23 +8,11 @@ import (
 )
 
 func main() {
-	// key := 1
-	// value := []byte("hello friend")
 
 	t := bptree.NewTree()
 
-	// err := t.Insert(key, value)
-	// if err != nil {
-	// 	fmt.Printf("error: %s\n\n", err)
-	// }
-
-	// r, err := t.Find(key, true)
-	// if err != nil {
-	// 	fmt.Printf("error: %s\n\n", err)
-	// }
-
-	// fmt.Printf("%s\n\n", r.Value)
 	// Insert 1, 4, 7, 10, 17, 21, 31, 25, 19, 20, 28, 42
+	// lecture example
 
 	arr := [12]int{1, 4, 7, 10, 17, 21, 31, 25, 19, 20, 28, 42}
 
@@ -36,5 +24,12 @@ func main() {
 	}
 
 	t.PrintTree()
+
+	r, err := t.Find(7, true)
+	if err != nil {
+		fmt.Printf("error: %s\n\n", err)
+	}
+
+	fmt.Printf("%s\n\n", r.Value)
 
 }
